+++ src/handlers/all-sectors/report/releases.js
@@ -1,502 +1,170 @@
 'use strict';
 
-/**
- * Common functions for route handlers for substance releases to
- * air, water, waste water and land
- */
 const MasterDataService = require('../../../service/master-data');
-const Validator = require('../../../lib/validator');
-const CacheKeyError = require('../../../lib/user-cache-policies').CacheKeyError;
-const errHdlr = require('../../../lib/utils').generalErrorHandler;
 
-const cacheHelper = require('../common').cacheHelper;
 const isNumeric = require('../../../lib/utils').isNumeric;
 const cacheNames = require('../../../lib/user-cache-policies').names;
-
+const BaseStage = require('../common').BaseStage;
 const setConfirmation = require('../common').setConfirmation;
 const setValidationStatus = require('../common').setValidationStatus;
 const setChallengeStatus = require('../common').setChallengeStatus;
 
-const NEW_RELEASE_OBJECT = { value: null, unitId: null, methodId: null };
+const internals = {};
 
-const internals = {
-
-    /**
-     * Save submission to the task object regardless of validation status
-     * Does not write the task object back to the cache
-     * @param request
-     * @return {Promise.<void>}
-     */
-    save: (request, tasks) => {
-        if (tasks.releases) {
-            Object.keys(tasks.releases).filter(r => isNumeric(r)).forEach(s => {
-                // Strip the value from the payload and add to the tasks objects
-                tasks.releases[s].value = request.payload['value-' + s];
-
-                // Strip the units from the payload and add to the tasks objects
-                const unitId = Number.parseInt(request.payload['unitId-' + s]);
-                tasks.releases[s].unitId = Number.isNaN(unitId) ? null : unitId;
-            });
-        }
-    },
-
-    /**
-     * Validates the submission saving the state to the cache if the submission is invalid.
-     *
-     * Returns true if invalid
-     * @param request - the request object
-     * @param tasks - the tasks read from cache
-     * @return {Promise.<boolean>} - promises to be true if invalid
-     */
-    validate: (request, tasks) => {
-        if (tasks.releases) {
-            let isValid = true;
-            Object.keys(tasks.releases).filter(r => isNumeric(r)).forEach(s => {
-                delete tasks.releases[s].errors;
-                const validation = Validator.release(tasks.releases[s]);
-                if (validation) {
-                    tasks.releases[s].errors = validation;
-                    isValid = false;
-                }
-            });
-            return isValid;
-        }
-    },
-
-    /**
-     * Function to order the parameters on the release screen
-     * @param a - first release
-     * @param b - second release
-     * @return {number}
-     */
-    sortReleases: (a, b) => {
-        const nameA = a.parameter.nomenclature.toUpperCase();
-        const nameB = b.parameter.nomenclature.toUpperCase();
-
-        if (nameA < nameB) {
-            return -1;
-        }
-
-        if (nameA > nameB) {
-            return 1;
-        }
-
-        return 0;
-    },
-
-    /**
-     * Sort the substance list by name alphabetically
-     * @param a
-     * @param b
-     * @return {number}
-     */
-    sortParameters: (a, b) => {
-        const nameA = a.nomenclature.toUpperCase();
-        const nameB = b.nomenclature.toUpperCase();
-
-        if (nameA < nameB) {
-            return -1;
-        }
-
-        if (nameA > nameB) {
-            return 1;
-        }
-
-        return 0;
+class Confirm extends BaseStage {
+    constructor (...args) {
+        super(args);
     }
-};
+
+    async doGet (request, h, cacheState) {
+        if (cacheState.tasks && cacheState.tasks.releases && Object.keys(cacheState.tasks.releases)
+            .filter(r => isNumeric(r)).length > 0) {
+            // Redirect to main route page
+            return h.redirect(cacheState.route.page);
+
+        } else {
+            // Display the appropriate confirmation page
+            return h.view(this.path, {
+                route: cacheState.route,
+                selected: false
+            });
+        }
+    }
+
+    async doPost (request, h, cacheState) {
+        // Process the confirmation - the releases page
+        if (request.payload && request.payload.confirmation === 'true') {
+            await setChallengeStatus(request, cacheState.submissionContext, cacheState.route, true);
+            return h.redirect(cacheState.route.page + '/add-substance');
+        } else {
+        // If the challenge page results in false then this is a confirmed route
+            await setConfirmation(request, cacheState.submissionContext, cacheState.route, true);
+
+            // Unset the confirmation status when viewing the page
+            await setChallengeStatus(request, cacheState.submissionContext, cacheState.route);
+            return h.redirect('/task-list');
+        }
+    }
+}
+
+class Substances extends BaseStage {
+    constructor (...args) {
+        super(args);
+    }
+
+    async doGet (request, h, cacheState) {
+        const { route, tasks, submissionContext, eaId } = cacheState;
+        // Unset the confirmation status when viewing the page
+        await setConfirmation(request, submissionContext, route);
+
+        // Calculate the obligation for this route
+        const obligation = await MasterDataService.getReleaseObligation(eaId.regime.id, route);
+
+        // Remove any parameters already reported and sort the parameters and the groups
+        const parameterGroups = obligation.parameterGroups.map(g => {
+            const group = {};
+            group.id = g.id;
+            group.nomenclature = g.nomenclature;
+            if (tasks.releases) {
+                group.parameters = g.parameters.filter(p => !Object.keys(tasks.releases).map(n => Number.parseInt(n))
+                    .includes(p.id)).sort(internals.sortParameters);
+            } else {
+                group.parameters = g.parameters.sort(internals.sortParameters);
+            }
+            return group;
+        });
+
+        if (tasks.currentRelease && tasks.currentRelease.incomplete) {
+            return h.view(this.path, {
+                route: route,
+                parameterGroups: parameterGroups,
+                release: tasks.currentRelease.incomplete,
+                obligation: obligation
+            });
+        }
+
+        return h.view(this.path, {
+            route: route,
+            parameterGroups: parameterGroups,
+            obligation: obligation
+        });
+
+    }
+
+    async doPost (request, h, cacheState, errors) {
+        if (errors) {
+            cacheState.tasks.currentRelease = {
+                incomplete: {
+                    errors: errors,
+                    page: request.payload
+                }
+            };
+            await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, cacheState.tasks);
+            return h.redirect(cacheState.route.page + '/add-substance');
+        }
+        cacheState.tasks.currentRelease = {};
+        cacheState.tasks.currentRelease.parameterIds = request.payload.parameterId;
+        await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, cacheState.tasks);
+        return h.redirect(cacheState.route.page + '/details');
+    };
+}
+
+class Details extends BaseStage {
+    constructor (...args) {
+        super(args);
+    }
+
+    async doGet (request, h, cacheState) {
+        const { eaId, route, tasks } = cacheState;
+
+        const obligation = await MasterDataService.getReleaseObligation(eaId.regime.id, route);
+
+        const parameterMap = new Map([].concat(...obligation.parameterGroups
+            .map(g => g.parameters))
+            .map(p => [p.id, p]));
+
+        if (tasks.currentRelease.parameterIds && tasks.currentRelease.parameterIds.length) {
+            const release = {
+                parameter: parameterMap.get(Number.parseInt(tasks.currentRelease.parameterIds[0]))
+            };
+
+            obligation.units.sort((a, b) => a.conversion - b.conversion);
+
+            return h.view(this.path, {
+                route: route,
+                release: release,
+                obligation: obligation,
+                methods: await MasterDataService.getMethods()
+            });
+        }
+
+        return h.view(this.path);
+    }
+
+    async doPost (request, h, cacheState, errors) {
+        const { route } = cacheState;
+        return h.redirect(route.page + '/details');
+    }
+}
+
+internals.confirm = new Confirm('all-sectors/report/confirm');
+internals.substances = new Substances('all-sectors/report/add-substance', null, async (payload) => {
+    if (!payload.parameterId) {
+        return [ { key: 'parameter', errno: 'PI-1004' } ];
+    }
+});
+internals.details = new Details('all-sectors/report/release-details');
 
 module.exports = {
-
-    // Expose the tasks object
-    tasks: internals.tasks,
-
-    // Expose the validate function
-    validate: internals.validate,
-
-    // The substances also need to be sorted in the overseas waste
-    sortParameters: internals.sortParameters,
-
-    /**
-     * Display the confirmation pages - skip if the release type is pre-populated
-     * @param request
-     * @param task
-     * @param stageStatus - the
-     * @return {Promise.<boolean>}
-     */
     confirm: async (request, h) => {
-        try {
-            const { submissionContext, route, tasks } = await cacheHelper(request);
-
-            if (request.method === 'get') {
-
-                if (tasks && tasks.releases && Object.keys(tasks.releases).filter(r => isNumeric(r)).length > 0) {
-                    // Redirect to main route page
-                    return h.redirect(route.page);
-
-                } else {
-                    // Display the appropriate confirmation page
-                    return h.view('all-sectors/report/confirm', {
-                        route: route,
-                        selected: false
-                    });
-                }
-
-            } else {
-                // Process the confirmation - the releases page
-                if (request.payload && request.payload.confirmation === 'true') {
-                    await setChallengeStatus(request, submissionContext, route, true);
-                    return h.redirect(route.page);
-                } else {
-                    // If the challenge page results in false then this is a confirmed route
-                    await setConfirmation(request, submissionContext, route, true);
-
-                    // Unset the confirmation status when viewing the page
-                    await setChallengeStatus(request, submissionContext, route);
-                    return h.redirect('/task-list');
-                }
-            }
-        } catch (err) {
-            return errHdlr(err, h);
-        }
+        return internals.confirm.handler(request, h);
     },
 
-    /**
-     * Handler for the main releases submission pages to air, land, waste-water and controlled water
-     * @param request
-     * @param h
-     * @param task
-     * @return {Promise.<void>}
-     */
-    releases: async (request, h) => {
-        try {
-            const { submissionContext, route, eaId, tasks } = await cacheHelper(request);
-
-            if (tasks.releases && Object.keys(tasks.releases).filter(r => isNumeric(r)).length > 0) {
-                // Enrich the stored object for page presentation - add descriptions
-                let releases = [];
-
-                // Calculate the obligation for this route
-                const obligation = await MasterDataService.getReleaseObligation(eaId.regime.id, route);
-                const parameterMap = new Map([].concat(...obligation.parameterGroups
-                    .map(g => g.parameters))
-                    .map(p => [p.id, p]));
-
-                if (tasks.releases) {
-                    releases = Object.keys(tasks.releases).filter(r => isNumeric(r)).map(id => {
-                        return {
-                            parameter: (() => {
-                                const foundParameter = parameterMap.get(Number.parseInt(id));
-                                if (!foundParameter) {
-                                    throw new Error(`Unexpected parameter id: ${id} for obligation ${obligation.nomenclature}`);
-                                }
-                                return foundParameter;
-                            })(),
-                            value: tasks.releases[id].value,
-                            unitId: tasks.releases[id].unitId,
-                            errors: tasks.releases[id].errors
-                        };
-                    });
-
-                    // Sort the releases by substance name
-                    releases.sort(internals.sortReleases);
-                }
-
-                // Unset the confirmation status when viewing the page
-                await setConfirmation(request, submissionContext, route);
-
-                return h.view('all-sectors/report/releases', {
-                    route: route,
-                    eaId: eaId.nomenclature,
-                    releases: releases,
-                    units: obligation.units
-                });
-
-            } else {
-
-                // Add a release immediately
-                return h.redirect(route.page + '/add-substance');
-            }
-
-        } catch (err) {
-            return errHdlr(err, h);
-        }
+    substances: async (request, h) => {
+        return internals.substances.handler(request, h);
     },
 
-    /**
-     * Save action - responds to the post request on the main releases summary page
-     */
-    action: async (request, h) => {
-        try {
-            const { route, tasks, submissionContext } = await cacheHelper(request);
-
-            // Save the submission
-            await internals.save(request, tasks);
-
-            // If we continue we will need to validate the submission
-            if (request.payload.continue) {
-
-                // Test if the releases are valid
-                if (await internals.validate(request, tasks)) {
-                    // Set the confirmation flag to confirmed
-                    await setConfirmation(request, submissionContext, route, true);
-
-                    // Set the overall route validation status to valid
-                    await setValidationStatus(request, submissionContext, route, true);
-
-                    // Rewrite the tasks with no error and go back to the task-list
-                    await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                    return h.redirect('/task-list');
-                } else {
-                    // Unset the confirmation flag
-                    await setConfirmation(request, submissionContext, route);
-
-                    // Set the overall route validation status to invalid
-                    await setValidationStatus(request, submissionContext, route);
-
-                    // Rewrite the tasks and go back to the route page
-                    await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                    return h.redirect(route.page);
-                }
-
-            } else if (Object.keys(request.payload).find(s => s.startsWith('detail'))) {
-
-                // Save the parameter id and redirect to the release detail page
-                tasks.currentParameterId = Object.keys(request.payload)
-                    .find(s => s.startsWith('detail')).substr(7);
-
-                await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                return h.redirect(route.page + '/detail');
-
-            } else if (Object.keys(request.payload).find(s => s.startsWith('delete'))) {
-                // Save the parameter id and redirect to the delete confirmation page
-                const parameterId = Number.parseInt(Object.keys(request.payload)
-                    .find(s => s.startsWith('delete')).substr(7));
-
-                // Send to the delete confirmation dialog
-                tasks.currentParameterId = parameterId;
-                await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                return h.redirect(route.page + '/remove');
-
-            } else if (request.payload.add) {
-                // Save the release information to the cache and redirect to the add-substances page
-                await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                return h.redirect(route.page + '/add-substance');
-            }
-
-        } catch (err) {
-            return errHdlr(err, h);
-        }
-    },
-
-    /**
-     * Detail action
-     */
-    detail: async (request, h) => {
-        try {
-            // Check the permit status has been set
-            const { submissionContext, route, tasks, eaId } = await cacheHelper(request);
-
-            if (!tasks.releases || !tasks.currentParameterId) {
-                throw new CacheKeyError('Cache read error');
-            }
-
-            if (request.method === 'get') {
-                // Unset the confirmation status when viewing the page
-                await setConfirmation(request, submissionContext, route);
-
-                // Get the current release and enrich with the substance details
-                const release = tasks.releases[tasks.currentParameterId];
-
-                // Calculate the obligation for this route
-                const obligation = await MasterDataService.getReleaseObligation(eaId.regime.id, route);
-
-                // Create parameter map
-                const parameterMap = new Map([].concat(...obligation.parameterGroups
-                    .map(g => g.parameters))
-                    .map(p => [p.id, p]));
-
-                const parameter = parameterMap.get(Number.parseInt(tasks.currentParameterId));
-                release.parameter = parameter;
-
-                // Get the methods list
-                const methods = await MasterDataService.getMethods();
-
-                // Display the detail page
-                return h.view('all-sectors/report/release-detail', { route: route, release: release, methods: methods, units: obligation.units });
-            } else {
-
-                // Set the task detail elements
-                const { unitId, methodId, value, hasNotifiableRelease,
-                    notifiableUnitId, notifiableValue, notifiableReason } = request.payload;
-
-                const currentRelease = tasks.releases[tasks.currentParameterId];
-
-                // Set up the release object
-                currentRelease.unitId = Number.isNaN(Number.parseInt(unitId)) ? null : Number.parseInt(unitId);
-                currentRelease.methodId = Number.isNaN(Number.parseInt(methodId)) ? null : Number.parseInt(methodId);
-                currentRelease.value = Number.isNaN(Number.parseFloat(value)) ? value : Number.parseFloat(value);
-
-                if (hasNotifiableRelease === 'Yes') {
-                    currentRelease.notifiable = {};
-                    currentRelease.notifiable.unitId = Number.isNaN(Number.parseInt(notifiableUnitId)) ? null : Number.parseInt(notifiableUnitId);
-                    currentRelease.notifiable.value = Number.isNaN(Number.parseFloat(notifiableValue)) ? null : Number.parseFloat(notifiableValue);
-                    currentRelease.notifiable.reason = notifiableReason.substring(0, 500);
-                } else {
-                    delete currentRelease.notifiable;
-                }
-
-                delete currentRelease.errors;
-
-                // Validate the release object
-                const validation = Validator.release(tasks.releases[tasks.currentParameterId]);
-
-                if (validation) {
-                    // Unset the overall validation status
-                    await setValidationStatus(request, submissionContext, route);
-
-                    // Update the cache with the validation objects and redirect back to the releases page
-                    currentRelease.errors = validation;
-                    await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                    return h.redirect(route.page + '/detail');
-                } else {
-                    // Calculate the overall validation status
-                    await setValidationStatus(request, submissionContext, route, internals.validate(request, tasks));
-
-                    // Valid to go back to the main releases page
-                    await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                    return h.redirect(route.page);
-                }
-            }
-        } catch (err) {
-            return errHdlr(err, h);
-        }
-    },
-
-    /**
-     * Remove a release
-     * @param request
-     * @param h
-     * @return {Promise.<void>}
-     */
-    remove: async (request, h) => {
-        try {
-            const { route, tasks, submissionContext } = await cacheHelper(request);
-            const release = tasks.releases[tasks.currentParameterId];
-            const parameter = await MasterDataService.getParameterById(Number.parseInt(tasks.currentParameterId));
-
-            if (request.method === 'get') {
-                release.parameter = parameter;
-                return h.view('all-sectors/report/confirm-delete', { route: route, release: release });
-            } else {
-                delete tasks.releases[tasks.currentParameterId];
-                await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-
-                // Recalculate the overall route validation status
-                await setValidationStatus(request, submissionContext, route, internals.validate(request, tasks));
-
-                // If this is the last release redirect back to the task list
-                if (Object.keys(tasks.releases).filter(r => isNumeric(r)).length > 0) {
-                    return h.redirect(route.page);
-                } else {
-                    // Here we unset the challenge flag - the user must explicitly say no to the route
-                    await setChallengeStatus(request, submissionContext, route);
-                    return h.redirect('/task-list');
-                }
-            }
-        } catch (err) {
-            return errHdlr(err, h);
-        }
-    },
-
-    /**
-     * Add (single) substance
-     * @param request
-     * @param h
-     * @return {Promise.<void>}
-     */
-    add: async (request, h) => {
-        try {
-            // Get cache objects
-            const { route, tasks, submissionContext, eaId } = await cacheHelper(request);
-
-            if (request.method === 'get') {
-                // Unset the confirmation status when viewing the page
-                await setConfirmation(request, submissionContext, route);
-
-                // Calculate the obligation for this route
-                const obligation = await MasterDataService.getReleaseObligation(eaId.regime.id, route);
-
-                // Remove any parameters already reported and sort the parameters and the groups
-                const parameterGroups = obligation.parameterGroups.map(g => {
-                    const group = {};
-                    group.id = g.id;
-                    group.nomenclature = g.nomenclature;
-                    if (tasks.releases) {
-                        group.parameters = g.parameters.filter(p => !Object.keys(tasks.releases).map(n => Number.parseInt(n))
-                            .includes(p.id)).sort(internals.sortParameters);
-                    } else {
-                        group.parameters = g.parameters.sort(internals.sortParameters);
-                    }
-                    return group;
-                });
-
-                if (tasks.releases && tasks.releases.parameterErrors) {
-                    return h.view('all-sectors/report/add-substance', {
-                        route: route,
-                        parameterGroups: parameterGroups,
-                        errors: tasks.releases.parameterErrors
-                    });
-                } else {
-                    return h.view('all-sectors/report/add-substance', {
-                        route: route,
-                        parameterGroups: parameterGroups
-                    });
-                }
-
-            } else {
-                let success = false;
-
-                // Add a new releases array to the task object if it does not exist
-                if (!tasks.releases) {
-                    tasks.releases = {};
-                }
-
-                if (request.payload.parameterId) {
-                    const parameter = await MasterDataService.getParameterById(Number.parseInt(request.payload['parameterId']));
-
-                    // Add the selected parameters to the task if it exists
-                    if (parameter) {
-
-                        // Assign this parameter to the transfer object
-                        if (!tasks.releases[parameter.id]) {
-                            tasks.releases[parameter.id] = NEW_RELEASE_OBJECT;
-                        }
-
-                        // Immediately set the overall validation status to false
-                        await setValidationStatus(request, submissionContext, route);
-
-                        // Set the current task to allow us to get directly to the detail page
-                        tasks.currentParameterId = parameter.id;
-                        delete tasks.releases.parameterErrors;
-                        success = true;
-
-                        // If there is no parameter then redirect back with an error write the task object back to the cache
-                        await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-
-                        return h.redirect(route.page + '/detail');
-                    }
-                }
-
-                if (!success) {
-                    tasks.releases.parameterErrors = [ { key: 'parameter', errno: 'PI-1004' } ];
-                    await request.server.app.userCache.cache(cacheNames.TASK_CONTEXT).set(request, tasks);
-                    return h.redirect(route.page + '/add-substance');
-                }
-
-            }
-        } catch (err) {
-            return errHdlr(err, h);
-        }
+    details: async (request, h) => {
+        return internals.details.handler(request, h);
     }
 };
