+++ src/routes.js
@@ -6,7 +6,9 @@
 const Authentication = require('./handlers/authentication.js');
 const Start = require('./handlers/start.js');
 const AllSectors = require('./handlers/all-sectors/task-list');
+const ReleasesOld = require('./handlers/all-sectors/report/releases-old');
 const Releases = require('./handlers/all-sectors/report/releases');
+
 const Waste = require('./handlers/all-sectors/report/waste');
 const Review = require('./handlers/all-sectors/submit/review');
 const Submit = require('./handlers/all-sectors/submit/submit');
@@ -88,11 +90,11 @@ const dynamicHandlers = [
 
     // Releases to air, land, controlled waters and in waste-water
     { method: ['GET', 'POST'], path: '/releases/{route}/confirm', options: { handler: Releases.confirm } },
-    { method: 'GET', path: '/releases/{route}', options: { handler: Releases.releases } },
-    { method: 'POST', path: '/releases/{route}/action', options: { handler: Releases.action } },
-    { method: ['GET', 'POST'], path: '/releases/{route}/add-substance', options: { handler: Releases.add } },
-    { method: ['GET', 'POST'], path: '/releases/{route}/detail', options: { handler: Releases.detail } },
-    { method: ['GET', 'POST'], path: '/releases/{route}/remove', options: { handler: Releases.remove } },
+    { method: ['GET', 'POST'], path: '/releases/{route}/add-substance', options: { handler: Releases.substances } },
+    { method: ['GET', 'POST'], path: '/releases/{route}/details', options: { handler: Releases.details } },
+    // { method: 'GET', path: '/releases/{route}', options: { handler: ReleasesOld.releases } },
+    // { method: 'POST', path: '/releases/{route}/action', options: { handler: ReleasesOld.action } },
+    // { method: ['GET', 'POST'], path: '/releases/{route}/remove', options: { handler: ReleasesOld.remove } },
 
     // Waste Transfers
     { method: ['GET', 'POST'], path: '/transfers/waste/confirm', options: { handler: Waste.confirm } },
